Fix for trailing blank lines
If your document has trailing blank lines and Xcode is set to remove them on save then the file handed to clang-format can be shorter than the in-memory file and screw up how we determine how many files were added or deleted. The fix is to ignoring trailing blank lines - use the Xcode setting if you want the trimmed.